Clarify model evaluation explanation in notebook
This commit updates the explanation regarding the model's performance
evaluation in the Jupyter notebook for Lab 3. It clarifies why the model
is not considered overfitted by emphasizing the small difference in
accuracy between the training and testing sets. Additionally, it
highlights the resolution of the curse of dimensionality issue by
reducing the feature vector length from 128 to 16, further supporting
the model's effective learning without overfitting. The adjustment in
the description of the learning curve graph is also made to more
accurately reflect the model's learning process.